You can't always cast pointer to int on because the sizes are different; use intptr_t

+++ gnu/usr.sbin/mkhybrid/src/match.h
@@ -5,18 +5,19 @@
  */
 
 /*
- * 	$Id: match.h,v 1.1 2000/10/10 20:40:18 beck Exp $
+ * 	$Id: match.h,v 1.2 2002/11/03 20:27:03 pvalchev Exp $
  */
 
+#include <inttypes.h>
 #include "fnmatch.h"
 
 int matches	__PR((char *fn));
 
 int i_matches	__PR((char *fn));
-int i_ishidden	__PR((void));
+intptr_t i_ishidden	__PR((void));
 
 int j_matches	__PR((char *fn));
-int j_ishidden	__PR((void));
+intptr_t j_ishidden	__PR((void));
 
 #ifdef APPLE_HYB
 int add_match	__PR((char *fn));
@@ -25,7 +26,7 @@ int j_add_match __PR((char *fn));
 
 int hfs_add_match __PR((char *fn));
 int hfs_matches __PR((char *fn));
-int hfs_ishidden __PR((void));
+intptr_t hfs_ishidden __PR((void));
 
 void add_list __PR((char *fn));
 void i_add_list __PR((char *fn));
